geocoding_service/requirements.txt
Flask==3.1.0
Flask-RESTful==0.3.10
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
flask-sqlalchemy==3.1.1
kafka-python==2.1.5
requests==2.31.0
geopy==2.4.1
haversine==2.8.1
python-dotenv==1.0.1
cachetools==5.3.3